2_13: OTA size u16->u32 (dong bo webtools/2_13inch)
Firmware > 64KB (v1.0/v1.1) lam u16 tran -> brick. Doi setUint16->setUint32 tai
buf+2 (tuong thich nguoc). Push nhanh 2-13inch (KHONG main - main go-live).

+++ js/2_13/main.js
@@ -492,7 +492,9 @@ async function otaUpdate() {
     const buf = new Uint8Array(136);
     const dv = new DataView(buf.buffer);
     buf[0] = 0xa0; buf[1] = 0x00;
-    dv.setUint16(2, firmSize, true);
+    // size u32 (truoc u16): firmware > 64KB. Tuong thich nguoc: fw cu doc u16
+    // (2 byte thap) van dung khi firmware dich < 64KB (vd ban "temp" cau noi).
+    dv.setUint32(2, firmSize, true);
     show('Đang xoá flash…');
     await write(buf, true);
